% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumsum_series.R
\name{cumsum_series}
\alias{cumsum_series}
\title{Create a cumulative sum series}
\usage{
cumsum_series(x)
}
\arguments{
\item{x}{A numeric vector or time series.}
}
\value{
A numeric vector containing the cumulative sums of `x`.
}
\description{
Calculates the cumulative sum of a numeric series.
}
\examples{
x <- c(100, 8, 7, 8, 7)
cumsum_series(x)

}
