% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_cusum_series.R
\name{de_cumsum_series}
\alias{de_cumsum_series}
\title{Reconstruct a series from its cumulative sum}
\usage{
de_cumsum_series(x_cum)
}
\arguments{
\item{x_cum}{A numeric vector containing cumulative sums, usually
created using `cumsum_series()`.}
}
\value{
A numeric vector with the same length as `x_cum`
  containing the reconstructed original series.
}
\description{
Reconstructs the original series from a cumulative sum series.
}
\examples{
x <- c(100, 8, 7, 8, 7)
x_cum <- cumsum_series(x)

de_cumsum_series(x_cum)

}
